refactor: encapsulate state root computation in StateRootCalculator
Replace free function and manual state threading with a StateRootCalculator
struct that owns cached trie updates and cumulative prefix sets internally.
FlashblocksState always holds a calculator; incremental vs full mode is
configured at construction. Move tests into state_root module and add e2e test.
