+++ crates/op-rbuilder/src/builder/payload.rs
@@ -1,4 +1,4 @@
-use super::wspub::WebSocketPublisher;
+use super::{state_root::StateRootCalculator, wspub::WebSocketPublisher};
 use crate::{
     backrun_bundle::BackrunBundlesPayloadCtx,
     builder::{
@@ -46,7 +46,7 @@ use reth_revm::{
     State, database::StateProviderDatabase, db::states::bundle_state::BundleRetention,
 };
 use reth_transaction_pool::TransactionPool;
-use reth_trie::{HashedPostState, prefix_set::TriePrefixSetsMut, updates::TrieUpdates};
+use reth_trie::{HashedPostState, updates::TrieUpdates};
 use revm::Database;
 use std::{collections::BTreeMap, sync::Arc, time::Instant};
 use tokio::sync::mpsc;
@@ -104,21 +104,12 @@ pub(super) struct FlashblocksState {
     da_footprint_per_batch: Option<u64>,
     /// Whether to disable state root calculation for each flashblock
     disable_state_root: bool,
-    /// Whether to enable incremental state root calculation using cached trie nodes
-    enable_incremental_state_root: bool,
     /// Index into ExecutionInfo tracking the last consumed flashblock
     /// Used for slicing transactions/receipts per flashblock
     last_flashblock_tx_index: usize,
-    /// Cached trie updates from previous flashblock for incremental state root calculation.
-    /// None only for the first flashblock; populated after each subsequent state root calculation.
-    prev_trie_updates: Option<Arc<TrieUpdates>>,
-    /// Cumulative prefix sets from all previous flashblocks in this block.
-    /// Extended into the current flashblock's prefix sets so the trie walker re-visits
-    /// every path that was modified in earlier flashblocks. Without this, reverted storage
-    /// slots can leave stale cached hashes in the incremental trie (the walker skips
-    /// subtrees whose prefix isn't covered, using the cached hash which reflects the
-    /// pre-revert value).
-    cumulative_prefix_sets: Option<TriePrefixSetsMut>,
+    /// State root calculator. Manages cached trie updates and cumulative prefix
+    /// sets across flashblocks when incremental mode is enabled.
+    state_root_calculator: StateRootCalculator,
 }
 
 impl FlashblocksState {
@@ -130,7 +121,7 @@ impl FlashblocksState {
         Self {
             target_flashblock_count,
             disable_state_root,
-            enable_incremental_state_root,
+            state_root_calculator: StateRootCalculator::new(enable_incremental_state_root),
             ..Default::default()
         }
     }
@@ -152,10 +143,8 @@ impl FlashblocksState {
             da_per_batch: self.da_per_batch,
             da_footprint_per_batch: self.da_footprint_per_batch,
             disable_state_root: self.disable_state_root,
-            enable_incremental_state_root: self.enable_incremental_state_root,
             last_flashblock_tx_index: self.last_flashblock_tx_index,
-            prev_trie_updates: self.prev_trie_updates.clone(),
-            cumulative_prefix_sets: self.cumulative_prefix_sets.clone(),
+            state_root_calculator: self.state_root_calculator.clone(),
         }
     }
 
@@ -1046,7 +1035,7 @@ where
 pub(super) fn build_block<DB, P>(
     state: &mut State<DB>,
     ctx: &OpPayloadBuilderCtx,
-    fb_state: Option<&mut FlashblocksState>,
+    mut fb_state: Option<&mut FlashblocksState>,
     info: &mut ExecutionInfo,
     calculate_state_root: bool,
     enable_tx_tracking_debug_logs: bool,
@@ -1102,7 +1091,6 @@ where
     let mut state_root = B256::ZERO;
     let mut hashed_state = HashedPostState::default();
     let mut trie_updates_to_cache: Option<Arc<TrieUpdates>> = None;
-    let mut prefix_sets_to_cache: Option<TriePrefixSetsMut> = None;
 
     let flashblock_index_for_trace = fb_state
         .as_deref()
@@ -1112,15 +1100,6 @@ where
     if calculate_state_root {
         let state_provider = state.database.as_ref();
 
-        let enable_incremental = fb_state
-            .as_deref()
-            .is_some_and(|s| s.enable_incremental_state_root);
-        let prev_trie = fb_state
-            .as_deref()
-            .and_then(|s| s.prev_trie_updates.clone());
-        let prev_cumulative_prefix_sets = fb_state
-            .as_deref()
-            .and_then(|s| s.cumulative_prefix_sets.clone());
         let flashblock_index = fb_state
             .as_deref()
             .map(|s| s.flashblock_index())
@@ -1128,33 +1107,31 @@ where
 
         hashed_state = state_provider.hashed_post_state(&state.bundle_state);
 
+        let calc = match fb_state.as_deref_mut() {
+            Some(s) => &mut s.state_root_calculator,
+            None => &mut StateRootCalculator::default(),
+        };
+
         debug!(
             target: "payload_builder",
             flashblock_index,
-            incremental = enable_incremental && prev_trie.is_some(),
+            incremental = calc.has_cached_trie(),
             "Computing state root"
         );
 
-        let result = super::state_root::compute_state_root(
-            state_provider,
-            hashed_state.clone(),
-            prev_trie.as_deref(),
-            prev_cumulative_prefix_sets,
-            enable_incremental,
-        )
-        .inspect_err(|err| {
-            warn!(
-                target: "payload_builder",
-                parent_header=%ctx.parent().hash(),
-                %err,
-                "failed to calculate state root for payload"
-            );
-        })
-        .map_err(PayloadBuilderError::other)?;
-
-        state_root = result.state_root;
-        prefix_sets_to_cache = Some(result.cumulative_prefix_sets);
-        trie_updates_to_cache = Some(Arc::new(result.trie_updates));
+        let output = calc
+            .compute(state_provider, hashed_state.clone())
+            .inspect_err(|err| {
+                warn!(
+                    target: "payload_builder",
+                    parent_header=%ctx.parent().hash(),
+                    %err,
+                    "failed to calculate state root for payload"
+                );
+            })
+            .map_err(PayloadBuilderError::other)?;
+        state_root = output.state_root;
+        trie_updates_to_cache = Some(output.trie_updates);
 
         let state_root_calculation_time = state_root_start_time.elapsed();
         ctx.metrics
@@ -1316,10 +1293,6 @@ where
 
     // pick the new transactions from the info field and update the last flashblock index
     let (new_transactions, new_receipts) = if let Some(fb_state) = fb_state {
-        if let Some(updates) = trie_updates_to_cache.take() {
-            fb_state.prev_trie_updates = Some(updates);
-        }
-        fb_state.cumulative_prefix_sets = prefix_sets_to_cache;
         let new_txs = fb_state.slice_new_transactions(&info.executed_transactions);
         let new_receipts = fb_state.slice_new_receipts(&info.receipts);
         fb_state.set_last_flashblock_tx_index(info.executed_transactions.len());
